requirements.txt
Flask==2.3.3
flask-wtf==1.1.1
wtforms==3.0.1
pandas==2.3.2
openpyxl==3.1.2
supabase==1.0.4
python-dotenv==1.0.0
werkzeug==2.3.7
